+++ src/App.vue
@@ -2,12 +2,12 @@
 <script setup lang="ts">
 import { RouterView, useRouter } from 'vue-router'
 import { useInterfaceStore } from './stores/interface'
-import { useWorkspacesStore } from './stores/workspaces'
+import { useWorkspacesStore } from './stores/collections'
 import { useSpacesStore } from './stores/spaces'
 import ItemEditorPanel from './components/items/ItemEditorPanel.vue'
 import ItemViewerPanel from './components/items/ItemViewerPanel.vue'
-import WorkspaceEditorPanel from './components/workspaces/WorkspaceEditorPanel.vue'
-import WorkspacesPanel from './components/workspaces/WorkspacesPanel.vue'
+import WorkspaceEditorPanel from './components/collections/CollectionEditorPanel.vue'
+import WorkspacesPanel from './components/collections/CollectionsPanel.vue'
 import SpaceEditorPanel from './components/spaces/SpaceEditorPanel.vue'
 import SpacesPanel from './components/spaces/SpacesPanel.vue'
 import SpeedbumpDialog from './SpeedbumpDialog.vue'
