+++ src/components/collections/CollectionChip.vue
@@ -1,7 +1,7 @@
 <!-- Renders a small colored chip for a workspace, adding a subtle border when the workspace color is very light. -->
 <script setup lang="ts">
 import { computed } from 'vue'
-import { useWorkspacesStore } from '@/stores/workspaces'
+import { useWorkspacesStore } from '@/stores/collections'
 import { isLightColor } from '@/utils/colors'
 
 const props = defineProps<{
