+++ src/main.ts
@@ -7,7 +7,7 @@ import { createPinia } from 'pinia'
 import App from './App.vue'
 import router from './router'
 import vuetify from './plugins/vuetify'
-import { useWorkspacesStore } from './stores/workspaces'
+import { useWorkspacesStore } from './stores/collections'
 import { useItemsStore } from './stores/items'
 import { useInterfaceStore } from './stores/interface'
 import { useSpacesStore } from './stores/spaces'
