+++ src/stores/spaces/scrubbing.ts
@@ -2,7 +2,7 @@
 import { putSpace } from '@/db'
 import type { Space } from '@/types/spaces'
 import type { Ref } from 'vue'
-import { useWorkspacesStore } from '../workspaces'
+import { useWorkspacesStore } from '../collections'
 import { useVisualizationsStore } from '../visualizations'
 
 export function useSpacesScrubbing(
