+++ src/views/SpaceView.vue
@@ -4,7 +4,7 @@ import { computed, onUnmounted, ref, watch } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { useSpacesStore } from '@/stores/spaces'
 import { useVisualizationsStore } from '@/stores/visualizations'
-import { useWorkspacesStore } from '@/stores/workspaces'
+import { useWorkspacesStore } from '@/stores/collections'
 import { useInterfaceStore } from '@/stores/interface'
 import { constructNewVisualization, generateVisualizationId } from '@/types/visualizations'
 import RoadmapPane from '@/components/roadmap/RoadmapPane.vue'
