+++ src/views/TestView.vue
@@ -3,10 +3,10 @@
 import { computed, ref } from 'vue'
 import ItemEditorPanel from '@/components/items/ItemEditorPanel.vue'
 import ItemViewerPanel from '@/components/items/ItemViewerPanel.vue'
-import WorkspaceEditorPanel from '@/components/workspaces/WorkspaceEditorPanel.vue'
-import WorkspacesPanel from '@/components/workspaces/WorkspacesPanel.vue'
+import WorkspaceEditorPanel from '@/components/collections/CollectionEditorPanel.vue'
+import WorkspacesPanel from '@/components/collections/CollectionsPanel.vue'
 import RoadmapPane from '@/components/roadmap/RoadmapPane.vue'
-import { useWorkspacesStore } from '@/stores/workspaces'
+import { useWorkspacesStore } from '@/stores/collections'
 import { useInterfaceStore } from '@/stores/interface'
 import SpeedbumpDialog from '@/SpeedbumpDialog.vue'
 import BacklogList from '@/components/backlog/BacklogList.vue'
@@ -15,7 +15,7 @@ import { useVisualizationsStore } from '@/stores/visualizations'
 import { useSpacesStore } from '@/stores/spaces'
 import { useRouter } from 'vue-router'
 import { items as dummyItems } from '@/testing/dummy-items'
-import { workspaces as dummyWorkspaces } from '@/testing/dummy-workspaces'
+import { workspaces as dummyWorkspaces } from '@/testing/dummy-collections'
 import { visualizations as dummyVisualizations } from '@/testing/dummy-visualizations'
 import { spaces as dummySpaces } from '@/testing/dummy-spaces'
 import { clearDatabase } from '@/db'
